restructuring project to add tests, a config plan, and packaging for distribution
