+++ app/resources/fonts/darwin-pro-cufonfonts/Los Andes Type  Darwin Pro Regular It.otf
