+++ config/config.yml
@@ -1,0 +1,13 @@
+window_size: [600, 400]
+title: "Assinacraft"
+icon: "icon.png"
+font_primary: "resources/fonts/Montserrat/Montserrat-Regular.ttf"
+font_primary_size: 11
+font_primary_attr: ""
+font_secondary: "resources/fonts/Grandstander/Grandstander-Black.ttf"
+font_secondary_size: 15
+font_secondary_attr: "bold"
+theme: "DarkTeal12"
+key_codes:
+    QT_ENTER_KEY1: "special 16777220"
+    QT_ENTER_KEY2: "special 16777221"
