--- src/theme.yml
@@ -1,1 +1,0 @@
-theme: DarkTeal12
